Add Dice loss with whole thalamus
Add a new Dice loss object that also computes the Dice score of the
whole thalamus as well as individual labels. The total loss is then
a weighted average of both (1 - ) the whole thalamus Dice and the
individual label wise Dice scores.

In the previous version, which is still in the code, there appears to be
a bug where the Dice score computed for the loss was not calculated
label by label but rather as a "joint-Dice" metric slice-by-slice.

Simple modification also found that leaving the numerator of the Dice
calculation as it was would lead to missing labels always giving a 0
Dice score, even when the prediction was correct.