Add grouped Dice metric
Add the ability to specify a combination of Dice metrics. There are now
three Dice metrics that can be called to generate the loss.

The first is the original implementation by Eugenio, which looks at
label-wise Dice but may be inacurate when there are missing labels in a
training subject.

The second is DiceLossLabels, which mostly just adds a contriution from
the Dice for the whole thalamus. This additionally has a fix for the
missing labels problem (if a label was missing in the training labels
the dice for that channel would always be 1). The loss is then weighted
such that the whole thalamus dice is worth as much as the average dice
accross all labels.

The final loss is DiceLossGrouped. This is initialised with a group_seg
vector which assigns a group number to each channel in the onehot
encoded ground truth or softmax tensor. This can be set at run time for
training. group_seg is used with unsorted_segment_sum to allow for
calculation of Dice for the specified grouped labels. The loss is then
the average of three loss measures: the average label-level Dice lose,
the average group-level Dice loss, and the whole-thalamus-level Dice
loss.

This commit also reverts the mistaken "fix" for label-wise dice. It is
in-fact correct to use range(1, 4) to sum accross spatial dimensions as
an extra dimension is appended in the first dimension before training.