Add multi-seg training setup script
Add a script to start setting up the training set directory. This
creates a folder for each subject that contains a re-oriented and
cropped T1 as well as a folder containing the various Bayesian
segmentations of the thalamus.

Another piece of code is required to resample the available DTI in
the log-domain.