Fix zero dice for missing labels
In training subjects where one or more labels have no segmented voxels,
the loss would always be 1 even when the activations were sufficiently
low that no voxels would be segmented. We fix this by adding epsilon to
each voxel in the numerator as well as the denominator (to avoid Nans).