Add check for existing checkpoint - Issue #2
Adds a check of the model_dir to see if there are any existing wl2 or
Dice checkpoints. If there are and the checkpoint hasn't bee manually
set it will restart the training from the furthest checkpoint found.

If the name of the checkpoint matches the current loss name then the
training should pick up from the epoch after the one  specified in the
name.

Some testing may be required to check that the momentum, learning rate
etc. are maintained through this process.