Loop validate over subject list - Issue #3
Modify the validate function to loop over the whole subject list for a
single model instance. This allows the model to be reused for the whole
validation step rather than reloading it.

The loss is now calculated using the same metrics as used for training.
We no longer convert to hard segmentations and back but rather work on
the softmax activations.